tests/wpt/web-platform-tests/hr-time/monotonic-clock.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>window.performance.now() chronology</title>
<link rel="author" title="W3C" href="http://www.w3.org/" />
<link rel="help" href="http://www.w3.org/TR/hr-time/#sec-monotonic-clock"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(window.performance.now() > 0, "window.performance.now() returns positive numbers");
}, "window.performance.now() returns a positive number", {assert: "The time values returned when calling the now method MUST be monotonically increasing and not subject to system clock adjustments or system clock skew."});

test(function() {
    var now1 = window.performance.now();
    var now2 = window.performance.now();
    assert_true((now2-now1) >= 0, "window.performance.now() difference is not negative");
  },
  "window.performance.now() difference is not negative",
  {
    assert: "The difference between any two chronologically recorded time values returned from the now method MUST never be negative."
  }
);

</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that window.performance.now() follows the monotonic clock requirements.</p>

<div id="log"></div>

</body>
</html>
