third_party/WebKit/LayoutTests/platform/win/external/wpt/hr-time/monotonic-clock-expected.txt

This is a testharness.js-based test.
PASS window.performance.now() returns a positive number 
PASS window.performance.now() difference is not negative 
Harness: the test ran to completion.
